Update and rename snake game.py to snake game updated 1.01.py
 my simple snake game with difficulty selector and game over message and arrow key controls for movement
